Consolidate bounding functions for faces and elements into a new pair of source files
