+++ include/xdg/primitive_ref.h
@@ -12,7 +12,6 @@ struct PrimitiveRef {
 };
 
 void TriangleIntersectionFunc(RTCIntersectFunctionNArguments* args);
-void TriangleBoundsFunc(RTCBoundsFunctionArguments* args);
 void TriangleOcclusionFunc(RTCOccludedFunctionNArguments* args);
 bool TriangleClosestFunc(RTCPointQueryFunctionArguments* args);
